refactor: fold support script package and share test runner
Fold @ericsanchezok/synergy-script into packages/synergy/script/script-identity.ts
next to its sole consumer (build.ts) and drop the packages/script workspace
package, knip entry, root devDependency, and coverage-exempt entry. Version
derivation now calls the canonical script/release/shared/runtime.ts helpers.

Extract the duplicated sharded test runner from packages/app and packages/ui
into script/shared/test-runner.ts with parameterized per-package test lists.
