--- packages/script/package.json
@@ -1,11 +1,0 @@
-{
-  "$schema": "https://json.schemastore.org/package",
-  "name": "@ericsanchezok/synergy-script",
-  "license": "MIT",
-  "devDependencies": {
-    "@types/bun": "catalog:"
-  },
-  "exports": {
-    ".": "./src/index.ts"
-  }
-}
