+++ packages/synergy/package.json
@@ -29,7 +29,6 @@
     "@ast-grep/cli-win32-x64-msvc": "0.40.5",
     "@ericsanchezok/synergy-link": "workspace:*",
     "@ericsanchezok/synergy-plugin-kit": "workspace:*",
-    "@ericsanchezok/synergy-script": "workspace:*",
     "@parcel/watcher-darwin-arm64": "2.5.6",
     "@parcel/watcher-darwin-x64": "2.5.6",
     "@parcel/watcher-linux-arm64-glibc": "2.5.6",
@@ -73,7 +72,6 @@
     "@ericsanchezok/synergy-browser": "workspace:*",
     "@ericsanchezok/synergy-link-protocol": "workspace:*",
     "@ericsanchezok/synergy-plugin": "workspace:*",
-    "@ericsanchezok/synergy-script": "workspace:*",
     "@ericsanchezok/synergy-sdk": "workspace:*",
     "@ericsanchezok/synergy-util": "workspace:*",
     "@hono/zod-validator": "catalog:",
