+++ packages/synergy/script/build.ts
@@ -13,7 +13,7 @@ const dir = path.resolve(__dirname, "..")
 process.chdir(dir)
 
 import pkg from "../package.json"
-import { Script } from "@ericsanchezok/synergy-script"
+import { Script } from "./script-identity"
 import {
   assertPackagedSandboxAsset,
   copySandboxAsset,
